+++ src/storage/index.ts
@@ -1,3 +1,3 @@
 export * from './sync-record';
 export * from './store.interface';
-export * from './indexeddb-sync-state-store';
+export * from './store';